Prompts refactor
